[Rename] user.jsx 파일 이름 userAPI로 변경
